friend exception 70%

+++ src/main/java/org/sparta/newsfeed/friend/service/FriendService.java
@@ -22,69 +22,84 @@ public class FriendService {
     private final FriendRepository friendRepository;
     private final UserService userService;
 
+    /*유저가 친구 요청을 하는 메서드*/
     public void requestFriends(AuthUser authUser, FriendDto friendDto) {
-        User user = userService.findUserByEmail(authUser.getEmail());
-        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail());
+        User user = userService.findUserByEmail(authUser.getEmail()); // 보낸 사람의 userEntity를 가져옴
+        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail()); // 받는 사람의 userEntity를 가져옴
 
 
-
+        /*예외처리*/
         if (user == null || requestUser == null) {
             throw new IllegalArgumentException("유저를 찾지 못했습니다.");
         }
         if (user.equals(requestUser)){
             throw new IllegalArgumentException("자신에게 친구 요청을 할 수 없습니다.");
         }
-        Friend isAlreadyFriends = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElse(null);
+        Friend isAlreadyFriends = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElse(null); // Baseid와 FriendId가 들어온 값과 똑같은 Friend객체 가져옴 즉 친구요청을 했거나 친구인 상태인지 검증하기 위함
         if(isAlreadyFriends != null) {
             if (isAlreadyFriends.isApplyYn()) {
-                throw new IllegalArgumentException("이미 친구입니다.");
+                throw new IllegalArgumentException("이미 친구입니다."); 
             } else if (isAlreadyFriends.isApplyYn() == false) {
                 throw new IllegalArgumentException("이미 친구 요청을 하였습니다.");
             }
         }
+        else {
+            Friend friend = new Friend(requestUser,user, false); // 새로운 Friend객체 생성, applyYn = fals은 요청 대기상태
+            friendRepository.save(friend);
+        }
 
 
 
 
-        Friend friend = new Friend(user, requestUser, false);
-        Friend reverseFriend = new Friend(requestUser, user, false);
-        friendRepository.save(friend);
-        friendRepository.save(reverseFriend);
 
     }
 
+    /*유저가 친구 요청을 반려하는 메서드*/
     public void cancelRequestFriends(AuthUser authUser, FriendDto friendDto) {
-        User user = userService.findUserByEmail(authUser.getEmail());
-        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail());
+        User user = userService.findUserByEmail(authUser.getEmail()); // 보낸사람 의 UserEntity 가져옴
+        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail()); // 받는 사람의 UserEntity 가져옴
 
+        /*예외 처리*/
         if (user == null || requestUser == null) {
             throw new IllegalArgumentException("유저를 찾지 못했습니다.");
         }
-
-        Friend friend = friendRepository.findByBaseIdAndFriendId(user, requestUser).orElseThrow
+        if (user.equals(requestUser)) {
+            throw new IllegalArgumentException("자신에게 친구 요청을 할 수 없습니다.");
+        }
+        Friend friend = friendRepository.findByBaseIdAndFriendId(requestUser, user).orElseThrow
                 (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
-        Friend reversefriend = friendRepository.findByBaseIdAndFriendId(requestUser, user).orElseThrow
-                (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
+        if(/*거짓*/!friend.isApplyYn())
+        {
+            friendRepository.delete(friend);
+        }
+        else
+        {
+            throw new RuntimeException("이미 친구입니다.");
+        }
 
-        friendRepository.delete(friend);
-        friendRepository.delete(reversefriend);
+
     }
 
     public void deleteFriends(AuthUser authUser, FriendDto friendDto) {
-        User user = userService.findUserByEmail(authUser.getEmail());
-        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail());
+        User user = userService.findUserByEmail(authUser.getEmail()); //보낸사람의 UserEntity
+        User requestUser = userService.findUserByEmail(friendDto.getRequestEmail()); // 받은사람의 UserEntity 
 
+        /*예외 처리*/
         if (user == null || requestUser == null) {
-            throw new IllegalArgumentException("유저를 찾지 못했습니다.");
+            throw new IllegalArgumentException("유저를 찾지 못했습니다."); 
         }
         Friend friend = friendRepository.findByBaseIdAndFriendId(user, requestUser).orElseThrow
-                (() -> new RuntimeException("친구가 아닙니다.213"));
+                (() -> new RuntimeException("친구가 아닙니다."));
+        Friend reverseFriend = friendRepository.findByBaseIdAndFriendId(requestUser, user).orElseThrow
+                (() -> new RuntimeException("친구가 아닙니다."));
+
 
         if (friend.isApplyYn()) {
             friendRepository.delete(friend);
-            Friend reverseFriend = friendRepository.findByBaseIdAndFriendId(requestUser, user).orElseThrow
-                    (() -> new RuntimeException("친구가 아닙니다."));
             friendRepository.delete(reverseFriend);
+        }
+        else {
+            throw new RuntimeException("두 유저는 친구가 아닙니다.");
         }
 
     }
@@ -105,10 +120,16 @@ public class FriendService {
                 friend1.getBaseId().getName(), friend1.getBaseId().getCreatedAt())).toList();
     }
 
+    /*유저가 친구 요청을 보낸 리스트를 갖고옴*/
     public List<FriendResponseDto> getRequestFriends(AuthUser authUser) {
         List<Friend> requestFriends;
         User user = userService.findUserByEmail(authUser.getEmail());
         requestFriends = friendRepository.findByFriendIdAndApplyYnFalse(user);
+
+        if(requestFriends.isEmpty()) {
+            throw new RuntimeException("신청한 친구요청이 없습니다.");
+        }
+
 
 
         return requestFriends.stream().map(friend -> new FriendResponseDto(
@@ -122,16 +143,23 @@ public class FriendService {
         if (user == null || requestUser == null) {
             throw new IllegalArgumentException("유저를 찾지 못했습니다.");
         }
+        Friend isAlreadyFriends = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElse(null);
+        if(user.equals(requestUser)) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+        if(isAlreadyFriends.isApplyYn())
+        {
+            throw new RuntimeException("이미 친구입니다.");
+        }
 
-        Friend friend = friendRepository.findByBaseIdAndFriendId(requestUser,user).orElseThrow
-                (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
-        Friend reverseFriend = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElseThrow
+        Friend friend = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElseThrow
                 (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
         friend.setApplyYn(true);
-        reverseFriend.setApplyYn(true);
-
         friendRepository.save(friend);
+
+        Friend reverseFriend = new Friend(requestUser, user, true);
         friendRepository.save(reverseFriend);
+
 
     }
 
@@ -142,14 +170,19 @@ public class FriendService {
         if (user == null || requestUser == null) {
             throw new IllegalArgumentException("유저를 찾지 못했습니다.");
         }
-
+        if(user.equals(requestUser)) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
         Friend friend = friendRepository.findByBaseIdAndFriendId(user, requestUser).orElseThrow
                 (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
-        Friend reverseFriend = friendRepository.findByBaseIdAndFriendId(user,requestUser).orElseThrow
-                (() -> new RuntimeException("친구요청을 찾지 못했습니다."));
+        if(/*거짓*/friend.isApplyYn()){
+            friendRepository.delete(friend);
+        }
+        else if(friend.isApplyYn())
+        {
+            throw new RuntimeException("이미 친구입니다.");
+        }
 
-        friendRepository.delete(friend);
-        friendRepository.save(reverseFriend);
     }
